fix : 내려줄 때 Double로 내려주도록 수정

+++ src/main/kotlin/com/detect_pothole/detect_pothole/domain/pothole/dto/PotholeRegistrationRequest.kt
@@ -5,8 +5,8 @@ import java.math.BigDecimal
 
 data class PotholeRegistrationRequest(
         val geotabId: Long,
-        val xacc: BigDecimal,
-        val yacc: BigDecimal,
-        val zacc: BigDecimal,
+        val xacc: Double,
+        val yacc: Double,
+        val zacc: Double,
         val point: Point
 )
